+++ packages/page-mixer/src/components/deposit/index.tsx
@@ -1,1 +1,1 @@
-export * from './DepositConsole';
+export * from './WithdrawConsole';
