+++ packages/page-mixer/src/components/withdraw/index.tsx
@@ -1,1 +1,1 @@
-export * from './WithdrawConsole';
+export * from './DepositConsole';
